[Journal] Adding persona check for free paper upload, consolidating make_user_verified test setup
